refactor: purge encoding/json from WASM build graph — TinyGo unlocked
Sequel to #2835 (foundational split of deployment + visorconfig).
This PR finishes the encoding/json elimination across every
transitive importer reachable from
pkg/skywireconfig/{autoconfigcmd, keypair, genvisor} under
GOOS=js GOARCH=wasm, which unblocks TinyGo compilation of the
install-page WASM.

Result: `tinygo build -target wasm -no-debug ./cmd/wasm` on
apt-repo now produces a 1.85 MB blob (vs 7.3 MB for the Go-WASM
build, ≈ 75% smaller).

Build-tag splits applied (everything moves to //go:build !js for
the json-using halves, plus a tiny js-side stub where the type
surface needs to remain visible):

  pkg/buildinfo/
    buildinfo.go              shared exports, no json
    buildinfo_native.go !js   init() that parses ldflags-injected
                              `go list -m -json` via json.Unmarshal
    buildinfo_js.go     js    init() that skips the json parse;
                              version info comes purely from
                              debug.ReadBuildInfo

  pkg/dmsg/disc/
    interface.go              EntryReader/Writer/APIClient (no json)
    interface_native.go !js   `var json = jsoniter.ConfigFastest`
    entry.go                  Entry struct + Validate/timestamps
    entry_native.go     !js   Sign + VerifySignature (json.Marshal
                              the canonical signed payload)
    fallback.go         !js   FallbackClient (imports pkg/logging)

  pkg/dmsgc/spec/
    spec.go                   types + non-json helpers
    spec_native.go      !js   DmsgConfig.MarshalJSON +
                              UnmarshalJSON (polymorphic single-
                              vs array-of-Deployments shape)

  pkg/logging/
    broadcaster.go      !js
    formatter.go        !js
    hooks.go            !js
    logger.go           !js
    logging.go          !js   all original code; logrus + json
                              transitive deps stay on the native
                              side.
    logging_js.go       js    minimal stub: MasterLogger / Logger
                              empty structs, NewMasterLogger +
                              PackageLogger + MustGetLogger +
                              SetOutputTo as no-op funcs. Just
                              enough surface for visorconfig.Common
                              and dmsg/disc to compile under js;
                              none of the methods get called from
                              the WASM-reachable code paths.

  pkg/routing/
    addr.go, packet.go,       remain untagged (no json/uuid; the
    failure.go,               primitive types — Port, RouteID,
    route_descriptor.go       RouteDescriptor — used by visorconfig)
    cascade.go          !js   imports uuid
    route.go            !js   imports encoding/json + uuid
    rule.go             !js   imports uuid (RuleHeaderSize stays
                              here; pkSize + routeDescriptorSize
                              moved to route_descriptor.go so the
                              untagged type def resolves)
    table.go            !js   imports uuid + logging

  pkg/visor/visorconfig/
    args.go                   appsList type def + shell-arg
                              parser (no json)
    args_native.go      !js   appsList.MarshalJSON +
                              UnmarshalJSON + unmarshalAppConfig
    types.go                  Duration type def, no json methods
    types_native.go     !js   Duration.MarshalJSON +
                              UnmarshalJSON
    parse.go            !js   imports blang/semver (transitive
                              json) — only invoked when loading
                              a config from disk
    read.go             !js   imports encoding/json + os file IO

Verification:
  GOOS=js GOARCH=wasm go list -deps \
    ./pkg/skywireconfig/{genvisor,autoconfigcmd,keypair} \
    | grep -E 'encoding/json|json-iterator|logrus|google/uuid|blang'
  → (empty)

  tinygo build -target wasm -no-debug ./cmd/wasm  (apt-repo)
  → 1,847,793 bytes (1.85 MB)

  go build ./...                                  — clean
  GOOS=js GOARCH=wasm go build ./pkg/skywireconfig/...
                                                   — clean
  go test on touched packages                     — pass
  golangci-lint on touched packages               — 0 issues

The Go-WASM build path is unchanged in behavior — apt-repo's
default `/generator/` install page still uses it. The TinyGo
variant at `/generator/tinygo/` now functions as a real smaller-
blob alternative (genvisor.MustMarshalJSON is stubbed under js,
so the "Download skywire.json" button shows a sentinel-error
JSON until a hand-rolled streaming V1 serializer lands as a
follow-up — tracked).
