+++ pkg/dmsg/disc/interface.go
@@ -17,16 +17,16 @@ package disc
 import (
 	"context"
 
-	jsoniter "github.com/json-iterator/go"
-
 	"github.com/skycoin/skywire/pkg/cipher"
 )
 
-// json is the package-wide JSON codec. Declared here (rather than
-// in client.go) because entry.go also uses it for Marshal/Unmarshal
-// and client.go is now build-tag-gated; the variable needs to be
-// available across all build configurations.
-var json = jsoniter.ConfigFastest
+// The package-wide `json` var (jsoniter.ConfigFastest) lives in
+// interface_native.go now — TinyGo's stdlib can't link encoding/json's
+// reflect runtime helpers, and jsoniter pulls encoding/json
+// transitively. Entry's Sign/VerifySignature also moved to a
+// !js-tagged file (entry_native.go); under js no code path in this
+// package calls Marshal/Unmarshal so the var is genuinely unused
+// there.
 
 // EntryReader provides read-only access to discovery entries.
 type EntryReader interface {
