+++ pkg/logging/logging.go
@@ -1,5 +1,9 @@
+//go:build !js
+
 /*
 Package logging provides application logging utilities
+
+Tagged off WASM — see logging_js.go for the build-tag rationale.
 */
 package logging
